% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{copy_font_files}
\alias{copy_font_files}
\title{Copy font files to destination directory}
\usage{
copy_font_files(font_files, dest_dir)
}
\arguments{
\item{font_files}{Character vector of paths to the source font files}

\item{dest_dir}{Path to the destination directory where fonts should be
installed}
}
\value{
Logical indicating whether all files were successfully copied
}
\description{
Copies font files to the system font directory
}
\details{
Attempts to copy each font file to the destination directory,
overwriting existing files if necessary. Returns FALSE if any file copy
operation fails.
}
\keyword{internal}
