% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{download_font}
\alias{download_font}
\title{Download font file}
\usage{
download_font(url, dest_path)
}
\arguments{
\item{url}{The direct download URL for the font file}

\item{dest_path}{File path where the font file should be saved}
}
\value{
Logical indicating whether the download was successful
}
\description{
Downloads a font file from a direct URL
}
\keyword{internal}
