% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{extract_font_files}
\alias{extract_font_files}
\title{Extract font files from zip archive}
\usage{
extract_font_files(zip_path, extract_dir)
}
\arguments{
\item{zip_path}{Path to the downloaded zip file}

\item{extract_dir}{Directory where the zip contents should be extracted}
}
\value{
Character vector of full paths to the extracted font files
}
\description{
Extracts .ttf and .otf files from a downloaded font zip archive
}
\details{
Searches recursively through the extracted zip contents to find all
font files, supporting both TrueType (.ttf) and OpenType (.otf) formats.
}
\keyword{internal}
