% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{font_hoist}
\alias{font_hoist}
\title{Make font families available in ggplot2}
\usage{
font_hoist(family_name, silent = FALSE)
}
\arguments{
\item{family_name}{name of the font family.}

\item{silent}{do you want to suppress the message?}
}
\value{
The function returns a message indicating whether the font was successfully loaded or not.
}
\description{
Sometimes, even when fonts are loaded on a computer, they are not immediately available for use in R. This function fixes that.
}
\details{
See \href{https://yjunechoe.github.io/posts/2021-06-24-setting-up-and-debugging-custom-fonts/}{June Choe's blog post explaining this}.
This function requires the user to have already downloaded the font on their computer, and to have set up \code{ragg} as described in the blog post.
}
\examples{
# load Roboto Condensed (must be loaded from Google Fonts on system already)
font_hoist("Roboto Condensed")

# try loading a font not on your computer, or that doesn't exist. It will give an informative error message.
font_hoist("Lobster Monster")
}
