% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{get_os}
\alias{get_os}
\title{Determine the current operating system}
\usage{
get_os()
}
\value{
A character string representing the operating system:
"windows" for Windows OS,
"macos" for macOS,
"linux" for Linux.
}
\description{
This function identifies the operating system on which R is running.
It returns a character string indicating the OS type, which can be
one of the following: "windows", "macos", or "linux".
}
\examples{
# Get the current operating system
os_type <- get_os()
print(os_type)
}
