% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{install_google_font}
\alias{install_google_font}
\title{Download and install a Google Font}
\usage{
install_google_font(
  font_name,
  api_key = "AIzaSyDOr3jWLtl4IP08yNaddV61_40f0YByPHo"
)
}
\arguments{
\item{font_name}{Character string of the font name (e.g. "Dancing Script")}

\item{api_key}{Google Fonts API key. Defaults to a hard-coded key that
has very generous usage limits. If you need your own key, you can get one
from the Google Cloud Console.}
}
\value{
Logical indicating whether the font was successfully installed
}
\description{
Downloads and installs a font family from Google Fonts using the
official API
}
