% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ti_2}
\alias{theme_ti_2}
\title{ggplot Theme version 2}
\usage{
theme_ti_2(
  base_size = 12,
  base_family = "Roboto Condensed",
  title_size = 18,
  subtitle_size = 12,
  axis_size = 10,
  caption_size = 8
)
}
\arguments{
\item{base_size}{Text Size}

\item{base_family}{Text Font}

\item{title_size}{Size of Title Text}

\item{subtitle_size}{Size of Subtitle Text}

\item{axis_size}{Size of Axis Text}

\item{caption_size}{Size of Caption Text}
}
\description{
ggplot Theme version 2
}
