% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ti_colors}
\alias{ti_colors}
\title{Set of colors used in the Teal Insights color palettes and theme}
\format{
An object of class \code{list} of length 20.
}
\usage{
ti_colors
}
\description{
Anchor colors used to create the TI color palettes. These can also be used as stand-alone colors
(e.g. ti_colors$gold returns the hex code \"#f8c03e\")
}
\keyword{datasets}
