% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ti_palettes}
\alias{ti_palettes}
\title{Set of color palettes used in the \code{scale_color_ti()} and \code{scale_fill_ti()}
functions}
\format{
An object of class \code{list} of length 9.
}
\usage{
ti_palettes
}
\description{
Palettes created using the TI anchor colors. To view the list of palettes and
their associated hex codes, use \code{ti_palettes}).
Note that the palettes used in plots interpolate between the anchor colors,
allowing each palette to be used with any number of colors (within reason!)
}
\keyword{datasets}
